+++ yarn-project/txe/src/oracle/txe_oracle.ts
@@ -32,6 +32,7 @@ import {
   computeContractClassId,
   computePoint,
   computePreaddress,
+  computeTaggingSecret,
   deriveKeys,
   getContractClassFromArtifact,
 } from '@aztec/circuits.js';
@@ -752,17 +753,8 @@ export class TXE implements TypedOracle {
 
   async getAppTaggingSecret(sender: AztecAddress, recipient: AztecAddress): Promise<Fr> {
     const senderCompleteAddress = await this.getCompleteAddress(sender);
-    const senderPreaddress = computePreaddress(
-      senderCompleteAddress.publicKeys.hash(),
-      senderCompleteAddress.partialAddress,
-    );
-    const ivskSender = await this.keyStore.getMasterIncomingViewingSecretKey(senderPreaddress);
-    // TODO: #8970 - Computation of address point from x coordinate might fail
-    const recipientAddressPoint = computePoint(recipient);
-    const curve = new Grumpkin();
-    // Given A (sender) -> B (recipient) and h == preaddress
-    // Compute shared secret as S = (h_A + ivsk_A) * Addr_Point_B
-    const sharedSecret = curve.mul(recipientAddressPoint, ivskSender.add(new Fq(senderPreaddress.toBigInt())));
+    const senderIvsk = await this.keyStore.getMasterIncomingViewingSecretKey(sender);
+    const sharedSecret = computeTaggingSecret(senderCompleteAddress, senderIvsk, recipient);
     // Silo the secret to the app so it can't be used to track other app's notes
     return poseidon2Hash([sharedSecret.x, sharedSecret.y, this.contractAddress]);
   }
